requirements.txt
Django
crispy_bootstrap5
django-crispy_forms
mysqlclient
django-tinymce
pillow



